backend/requirements.txt
# mPayHub backend dependencies (single source of truth)
#
# This file intentionally includes both runtime and development tooling
# so local/dev/prod installs can all use one command:
#   pip install -r requirements.txt

# Core Django and REST Framework
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.0

# Database
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0

# Security
django-cors-headers>=4.3.0
django-ratelimit>=4.1.0
cryptography>=41.0.0

# External Integrations
requests>=2.31.0

# Utilities
redis>=4.6.0
python-decouple>=3.8
pytz>=2023.3
Pillow>=10.0.0

# API Documentation
drf-spectacular>=0.26.0

# Production Server
gunicorn>=21.2.0
whitenoise>=6.6.0

# Development Tools
ipython>=8.17.0
django-debug-toolbar>=4.2.0
django-extensions>=3.2.0

# Code Quality
black>=23.11.0
flake8>=6.1.0
pylint>=3.0.0

# Testing
pytest>=7.4.3
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
